LCC function & move VectorCursor into xmtp_proto
