+++ xmtp_proto/src/traits.rs
@@ -18,8 +18,10 @@ pub(super) mod combinators;
 mod error;
 mod query;
 pub mod stream;
+mod vector_clock;
 pub use boxed_client::*;
 pub use error::*;
+pub use vector_clock::*;
 
 pub trait HasStats {
     fn aggregate_stats(&self) -> AggregateStats;
